python_mcp/requirements.txt
mcp>=1.13.1
starlette>=0.36.3
uvicorn>=0.30.6
python-jose>=3.3.0
httpx>=0.27.2
pdfminer.six>=20231228
pypdfium2>=4.30.0
Pillow>=10.3.0
pytesseract>=0.3.10
